feat: order claiming for owner :D

+++ src/main/java/net/refractored/joblistings/gui/MyOrders.kt
@@ -8,6 +8,8 @@ import net.kyori.adventure.text.format.TextDecoration
 import net.refractored.joblistings.JobListings.Companion.eco
 import net.refractored.joblistings.JobListings.Companion.spiGUI
 import net.refractored.joblistings.database.Database
+import net.refractored.joblistings.database.Database.Companion.orderDao
+import net.refractored.joblistings.gui.MyClaimedOrders.Companion.giveRefundableItems
 import net.refractored.joblistings.order.Order
 import net.refractored.joblistings.order.OrderStatus
 import net.refractored.joblistings.util.MessageUtil
@@ -186,7 +188,13 @@ class MyOrders {
                             )
                             gui.removeButton((index + 10) + (gui.currentPage * 45))
                             eco.depositPlayer(actor.player, (order.cost / 2) )
-                            Database.orderDao.delete(order)
+                            order.status = OrderStatus.INCOMPLETE
+                            // No need to refund items to assignee, if no items were ever turned in
+                            if (order.itemCompleted == 0){
+                                Database.orderDao.delete(order)
+                            } else {
+                                Database.orderDao.update(order)
+                            }
                             val assigneeMessage = Component.text()
                                 .append(MessageUtil.toComponent(
                                     "<green>One of your orders, <gray>"
@@ -217,6 +225,24 @@ class MyOrders {
                             event.whoClicked.sendMessage(
                                 MessageUtil.toComponent("<green>Order claimed!")
                             )
+                            val inventorySpaces = actor.player.inventory.storageContents.count{
+                                it == null || (it.isSimilar(order.item) && it.amount < it.maxStackSize)
+                            }
+                            if (inventorySpaces == 0) {
+                                actor.player.closeInventory()
+                                event.whoClicked.sendMessage(
+                                    MessageUtil.toComponent("<red>Your inventory is full!")
+                                )
+                                return@withListener
+                            }
+                            if (order.itemCompleted == order.itemsReturned) {
+                                actor.player.closeInventory()
+                                event.whoClicked.sendMessage(
+                                    MessageUtil.toComponent("<red>You have already refunded this order!")
+                                )
+                                return@withListener
+                            }
+                            giveOrderItems(order, actor)
                         }
 
                         OrderStatus.INCOMPLETE -> {
@@ -257,6 +283,42 @@ class MyOrders {
 
             }
             gui.refreshInventory(actor.player)
+        }
+
+        private fun giveOrderItems(order: Order ,actor: BukkitCommandActor){
+            val inventory = actor.player.inventory.storageContents
+            for ((inventoryIndex, inventoryItem) in inventory.withIndex()) {
+                if (order.itemCompleted == order.itemsObtained) break
+                if (inventoryItem == null) {
+                    val inventoryNewItem = order.item.clone().apply {
+                        amount = if (order.itemsObtained + maxStackSize >= order.itemCompleted) {
+                            order.itemCompleted - order.itemsObtained
+                        } else {
+                            maxStackSize
+                        }
+                    }
+                    order.itemsObtained += inventoryNewItem.amount
+                    actor.player.inventory.storageContents[inventoryIndex] = inventoryNewItem
+                    break
+                } else if (inventoryItem.isSimilar(order.item)) {
+                    if (inventoryItem.amount == inventoryItem.maxStackSize) continue
+                    val itemsNeeded = inventoryItem.maxStackSize - inventoryItem.amount
+                    val itemsToAdd = if (order.itemsObtained + itemsNeeded >= order.itemCompleted) {
+                        order.itemCompleted - order.itemsObtained
+                    } else {
+                        itemsNeeded
+                    }
+                    inventoryItem.amount += itemsToAdd
+                    order.itemsObtained += itemsToAdd
+                    break
+                }
+            }
+            orderDao.update(order)
+            if (order.itemsObtained == order.itemCompleted) {
+                MessageUtil.toComponent("<green>Order fully obtained!")
+                actor.player.closeInventory()
+                orderDao.delete(order)
+            }
         }
     }
 }
